Remove the proxy (un)install and upgrade for Kubernetes
From now on, the proxy on Kubernetes will be installed only using the
helm chart, so remove the corresponding code from the tools.

The support for Kubernetes is still kept for start, stop, restart.
cache clear, support config and logs for now.
