+++ mgradm/cmd/support/ptf/ptf_test.go
@@ -1,9 +1,9 @@
-// SPDX-FileCopyrightText: 2024 SUSE LLC
+// SPDX-FileCopyrightText: 2026 SUSE LLC
 //
 // SPDX-License-Identifier: Apache-2.0
 //go:build ptf
 
-package podman
+package ptf
 
 import (
 	"testing"
@@ -19,20 +19,23 @@ func TestParamsParsing(t *testing.T) {
 		"--ptf", "ptf123",
 		"--test", "test123",
 		"--user", "sccuser",
+		"--pullPolicy", "never",
+		"--registry", "myOldRegistry",
+		"--registry-host", "myoverwrittenregistry",
+		"--registry-user", "user",
+		"--registry-password", "password",
 	}
-
 	args = append(args, flagstests.SCCFlagTestArgs...)
-	args = append(args, flagstests.ImageProxyFlagsTestArgs...)
 
 	// Test function asserting that the args are properly parsed
-	tester := func(_ *types.GlobalFlags, flags *podmanPTFFlags,
-		_ *cobra.Command, _ []string,
-	) error {
-		flagstests.AssertSCCFlag(t, &flags.UpgradeFlags.SCC)
-		flagstests.AssertProxyImageFlags(t, &flags.UpgradeFlags.ProxyImageFlags)
+	tester := func(_ *types.GlobalFlags, flags *podmanPTFFlags, _ *cobra.Command, _ []string) error {
 		testutils.AssertEquals(t, "Error parsing --ptf", "ptf123", flags.PTFId)
 		testutils.AssertEquals(t, "Error parsing --test", "test123", flags.TestID)
 		testutils.AssertEquals(t, "Error parsing --user", "sccuser", flags.CustomerID)
+		testutils.AssertEquals(t, "Error parsing --pullPolicy", "never", flags.Image.PullPolicy)
+		testutils.AssertEquals(t, "Error parsing --registry", "myOldRegistry", flags.Image.Registry.Host)
+		flagstests.AssertRegistryFlag(t, &flags.Image.Registry)
+		flagstests.AssertSCCFlag(t, &flags.Installation.SCC)
 		return nil
 	}
