+++ mgrpxy/cmd/install/install.go
@@ -1,4 +1,4 @@
-// SPDX-FileCopyrightText: 2024 SUSE LLC
+// SPDX-FileCopyrightText: 2026 SUSE LLC
 //
 // SPDX-License-Identifier: Apache-2.0
 
@@ -6,22 +6,58 @@ package install
 
 import (
 	"github.com/spf13/cobra"
-	"github.com/uyuni-project/uyuni-tools/mgrpxy/cmd/install/kubernetes"
-	"github.com/uyuni-project/uyuni-tools/mgrpxy/cmd/install/podman"
+	"github.com/uyuni-project/uyuni-tools/mgrpxy/shared/podman"
+	"github.com/uyuni-project/uyuni-tools/mgrpxy/shared/utils"
 	. "github.com/uyuni-project/uyuni-tools/shared/l10n"
+	shared_podman "github.com/uyuni-project/uyuni-tools/shared/podman"
 	"github.com/uyuni-project/uyuni-tools/shared/types"
+	shared_utils "github.com/uyuni-project/uyuni-tools/shared/utils"
 )
 
-// NewCommand install a new proxy from scratch.
-func NewCommand(globalFlags *types.GlobalFlags) *cobra.Command {
-	installCmd := &cobra.Command{
+func newCmd(globalFlags *types.GlobalFlags, run shared_utils.CommandFunc[podman.PodmanProxyFlags]) *cobra.Command {
+	cmd := &cobra.Command{
 		Use:     "install [fqdn]",
+		Aliases: []string{"install podman"},
 		GroupID: "deploy",
-		Short:   L("Install a new proxy from scratch"),
-		Long:    L("Install a new proxy from scratch"),
-	}
-	installCmd.AddCommand(podman.NewCommand(globalFlags))
-	installCmd.AddCommand(kubernetes.NewCommand(globalFlags))
+		Short:   L("Install a new proxy on podman"),
+		Long: L(`Install a new proxy on podman
 
-	return installCmd
+It only takes the path to the configuration tarball generated by the server
+as parameter.
+
+The install podman command assumes podman is installed locally.
+
+/etc/uyuni/proxy/apache.conf and /etc/uyuni/squid.conf will be used as tuning files
+for apache and squid if available and not superseded by the matching command arguments.
+
+NOTE: for now installing on a remote podman is not supported!
+`),
+		Args: func(cmd *cobra.Command, args []string) error {
+			// ensure the right amount of args, managing podman
+			if len(args) > 0 && args[0] == "podman" {
+				return cobra.MaximumNArgs(2)(cmd, args)
+			}
+			return cobra.MaximumNArgs(1)(cmd, args)
+		},
+		RunE: func(cmd *cobra.Command, args []string) error {
+			// If the alias "install podman" is used, "podman" will be the first arg.
+			// We remove it from the args slice so it isn't treated as the FQDN.
+			if len(args) > 0 && args[0] == "podman" {
+				args = args[1:]
+			}
+			var flags podman.PodmanProxyFlags
+			return shared_utils.CommandHelper(globalFlags, cmd, args, &flags, nil, run)
+		},
+	}
+
+	utils.AddSCCFlag(cmd)
+	utils.AddImageFlags(cmd)
+	shared_podman.AddPodmanArgFlag(cmd)
+
+	return cmd
+}
+
+// NewCommand install a new proxy on podman from scratch.
+func NewCommand(globalFlags *types.GlobalFlags) *cobra.Command {
+	return newCmd(globalFlags, installForPodman)
 }
