+++ mgrpxy/cmd/install/install_test.go
@@ -1,31 +1,33 @@
-// SPDX-FileCopyrightText: 2025 SUSE LLC
+// SPDX-FileCopyrightText: 2026 SUSE LLC
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package kubernetes
+package install
 
 import (
+	"strings"
 	"testing"
 
 	"github.com/spf13/cobra"
-	"github.com/uyuni-project/uyuni-tools/mgrpxy/shared/kubernetes"
+	"github.com/uyuni-project/uyuni-tools/mgrpxy/shared/podman"
 	"github.com/uyuni-project/uyuni-tools/shared/testutils"
 	"github.com/uyuni-project/uyuni-tools/shared/testutils/flagstests"
 	"github.com/uyuni-project/uyuni-tools/shared/types"
 )
 
 func TestParamsParsing(t *testing.T) {
-	args := []string{}
-
+	args := []string{
+		"config.tar.gz",
+	}
 	args = append(args, flagstests.ImageProxyFlagsTestArgs...)
-	args = append(args, flagstests.ProxyHelmFlagsTestArgs...)
+	args = append(args, flagstests.PodmanFlagsTestArgs...)
+	args = append(args, flagstests.SCCFlagTestArgs...)
 
 	// Test function asserting that the args are properly parsed
-	tester := func(_ *types.GlobalFlags, flags *kubernetes.KubernetesProxyUpgradeFlags,
-		_ *cobra.Command, _ []string,
-	) error {
+	tester := func(_ *types.GlobalFlags, flags *podman.PodmanProxyFlags, _ *cobra.Command, _ []string) error {
 		flagstests.AssertProxyImageFlags(t, &flags.ProxyImageFlags)
-		flagstests.AssertProxyHelmFlags(t, &flags.Helm)
+		flagstests.AssertPodmanInstallFlags(t, &flags.Podman)
+		flagstests.AssertSCCFlag(t, &flags.SCC)
 		return nil
 	}
 
@@ -34,6 +36,7 @@ func TestParamsParsing(t *testing.T) {
 
 	testutils.AssertHasAllFlags(t, cmd, args)
 
+	t.Logf("flags: %s", strings.Join(args, " "))
 	cmd.SetArgs(args)
 	if err := cmd.Execute(); err != nil {
 		t.Errorf("command failed with error: %s", err)
