+++ mgrpxy/cmd/support/ptf/ptf.go
@@ -1,4 +1,4 @@
-// SPDX-FileCopyrightText: 2024 SUSE LLC
+// SPDX-FileCopyrightText: 2026 SUSE LLC
 //
 // SPDX-License-Identifier: Apache-2.0
 //go:build ptf
@@ -7,24 +7,53 @@ package ptf
 
 import (
 	"github.com/spf13/cobra"
-	"github.com/uyuni-project/uyuni-tools/mgrpxy/cmd/support/ptf/kubernetes"
-	"github.com/uyuni-project/uyuni-tools/mgrpxy/cmd/support/ptf/podman"
+	"github.com/uyuni-project/uyuni-tools/mgrpxy/shared/podman"
+	"github.com/uyuni-project/uyuni-tools/mgrpxy/shared/utils"
 	. "github.com/uyuni-project/uyuni-tools/shared/l10n"
 	"github.com/uyuni-project/uyuni-tools/shared/types"
+	shared_utils "github.com/uyuni-project/uyuni-tools/shared/utils"
 )
 
-// NewCommand is the command for creates supportptf.
+type podmanPTFFlags struct {
+	UpgradeFlags podman.PodmanProxyFlags `mapstructure:",squash"`
+	SCC          types.SCCCredentials    `mapstructure:"scc"`
+	PTFId        string                  `mapstructure:"ptf"`
+	TestID       string                  `mapstructure:"test"`
+	CustomerID   string                  `mapstructure:"user"`
+}
+
+func newCmd(globalFlags *types.GlobalFlags, run shared_utils.CommandFunc[podmanPTFFlags]) *cobra.Command {
+	var flags podmanPTFFlags
+	podmanCmd := &cobra.Command{
+		Use:     "ptf",
+		Aliases: []string{"ptf podman"},
+		Short:   L("Install a PTF or Test package on podman"),
+		Long: L(`Install a PTF or Test package on podman
+
+The support ptf podman command assumes podman is installed locally and
+the host machine is registered to SCC.
+
+NOTE: for now installing on a remote podman is not supported!
+`),
+		Args: func(cmd *cobra.Command, args []string) error {
+			// ensure the right amount of args, managing podman
+			if len(args) > 0 && args[0] == "podman" {
+				return cobra.MaximumNArgs(1)(cmd, args)
+			}
+			return cobra.MaximumNArgs(0)(cmd, args)
+		},
+		RunE: func(cmd *cobra.Command, args []string) error {
+			return shared_utils.CommandHelper(globalFlags, cmd, args, &flags, nil, run)
+		},
+	}
+
+	utils.AddSCCFlag(podmanCmd)
+	utils.AddImageFlags(podmanCmd)
+	shared_utils.AddPTFFlag(podmanCmd)
+	return podmanCmd
+}
+
+// NewCommand for podman installation.
 func NewCommand(globalFlags *types.GlobalFlags) *cobra.Command {
-	ptfCmd := &cobra.Command{
-		Use:   "ptf",
-		Short: L("Install a PTF"),
-	}
-
-	ptfCmd.AddCommand(podman.NewCommand(globalFlags))
-
-	if kubernetesCmd := kubernetes.NewCommand(globalFlags); kubernetesCmd != nil {
-		ptfCmd.AddCommand(kubernetesCmd)
-	}
-
-	return ptfCmd
+	return newCmd(globalFlags, ptfForPodman)
 }
