+++ mgrpxy/cmd/support/ptf/ptf_test.go
@@ -1,15 +1,14 @@
-// SPDX-FileCopyrightText: 2024 SUSE LLC
+// SPDX-FileCopyrightText: 2026 SUSE LLC
 //
 // SPDX-License-Identifier: Apache-2.0
+//go:build ptf
 
-package podman
+package ptf
 
 import (
-	"strings"
 	"testing"
 
 	"github.com/spf13/cobra"
-	"github.com/uyuni-project/uyuni-tools/mgrpxy/shared/podman"
 	"github.com/uyuni-project/uyuni-tools/shared/testutils"
 	"github.com/uyuni-project/uyuni-tools/shared/testutils/flagstests"
 	"github.com/uyuni-project/uyuni-tools/shared/types"
@@ -17,17 +16,23 @@ import (
 
 func TestParamsParsing(t *testing.T) {
 	args := []string{
-		"config.tar.gz",
+		"--ptf", "ptf123",
+		"--test", "test123",
+		"--user", "sccuser",
 	}
-	args = append(args, flagstests.ImageProxyFlagsTestArgs...)
-	args = append(args, flagstests.PodmanFlagsTestArgs...)
+
 	args = append(args, flagstests.SCCFlagTestArgs...)
+	args = append(args, flagstests.ImageProxyFlagsTestArgs...)
 
 	// Test function asserting that the args are properly parsed
-	tester := func(_ *types.GlobalFlags, flags *podman.PodmanProxyFlags, _ *cobra.Command, _ []string) error {
-		flagstests.AssertProxyImageFlags(t, &flags.ProxyImageFlags)
-		flagstests.AssertPodmanInstallFlags(t, &flags.Podman)
-		flagstests.AssertSCCFlag(t, &flags.SCC)
+	tester := func(_ *types.GlobalFlags, flags *podmanPTFFlags,
+		_ *cobra.Command, _ []string,
+	) error {
+		flagstests.AssertSCCFlag(t, &flags.UpgradeFlags.SCC)
+		flagstests.AssertProxyImageFlags(t, &flags.UpgradeFlags.ProxyImageFlags)
+		testutils.AssertEquals(t, "Error parsing --ptf", "ptf123", flags.PTFId)
+		testutils.AssertEquals(t, "Error parsing --test", "test123", flags.TestID)
+		testutils.AssertEquals(t, "Error parsing --user", "sccuser", flags.CustomerID)
 		return nil
 	}
 
@@ -36,7 +41,6 @@ func TestParamsParsing(t *testing.T) {
 
 	testutils.AssertHasAllFlags(t, cmd, args)
 
-	t.Logf("flags: %s", strings.Join(args, " "))
 	cmd.SetArgs(args)
 	if err := cmd.Execute(); err != nil {
 		t.Errorf("command failed with error: %s", err)
