+++ shared/kubernetes/kubernetes.go
@@ -5,25 +5,18 @@
 package kubernetes
 
 import (
-	"encoding/base64"
-	"fmt"
 	"os"
 	"strings"
 
 	"github.com/rs/zerolog"
 	"github.com/rs/zerolog/log"
 	. "github.com/uyuni-project/uyuni-tools/shared/l10n"
-	"github.com/uyuni-project/uyuni-tools/shared/types"
 	"github.com/uyuni-project/uyuni-tools/shared/utils"
-	core "k8s.io/api/core/v1"
-	meta "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime"
 )
 
 // ClusterInfos represent cluster information.
 type ClusterInfos struct {
 	KubeletVersion string
-	Ingress        string
 }
 
 // IsK3s is true if it's a K3s Cluster.
@@ -61,35 +54,11 @@ func CheckCluster() (*ClusterInfos, error) {
 
 	var infos ClusterInfos
 	infos.KubeletVersion = string(out)
-	infos.Ingress, err = guessIngress()
 	if err != nil {
 		return nil, err
 	}
 
 	return &infos, nil
-}
-
-func guessIngress() (string, error) {
-	// Check for a traefik resource
-	err := utils.RunCmd("kubectl", "explain", "ingressroutetcp")
-	if err == nil {
-		return "traefik", nil
-	}
-	log.Debug().Err(err).Msg("No ingressroutetcp resource deployed")
-
-	// Look for a pod running the nginx-ingress-controller: there is no other common way to find out
-	out, err := utils.RunCmdOutput(zerolog.DebugLevel, "kubectl", "get", "pod", "-A",
-		"-o", "jsonpath={range .items[*]}{.spec.containers[*].args[0]}{.spec.containers[*].command}{end}")
-	if err != nil {
-		return "", utils.Errorf(err, L("failed to get pod commands to look for nginx controller"))
-	}
-
-	const nginxController = "/nginx-ingress-controller"
-	if strings.Contains(string(out), nginxController) {
-		return "nginx", nil
-	}
-
-	return "", nil
 }
 
 // Restart restarts the pod.
@@ -113,123 +82,4 @@ func Start(namespace string, app string) error {
 // Stop stop the pod.
 func Stop(namespace string, app string) error {
 	return ReplicasTo(namespace, app, 0)
-}
-
-func get(component string, componentName string, args ...string) ([]byte, error) {
-	kubectlArgs := []string{
-		"get",
-		component,
-		componentName,
-	}
-
-	kubectlArgs = append(kubectlArgs, args...)
-
-	output, err := utils.RunCmdOutput(zerolog.DebugLevel, "kubectl", kubectlArgs...)
-	if err != nil {
-		return []byte{}, err
-	}
-	return output, nil
-}
-
-// GetConfigMap returns the value of a given config map.
-func GetConfigMap(configMapName string, filter string) (string, error) {
-	out, err := get("configMap", configMapName, filter)
-	if err != nil {
-		return "", utils.Errorf(err, L("failed to run kubectl get configMap %[1]s %[2]s"), configMapName, filter)
-	}
-
-	return string(out), nil
-}
-
-// GetSecret returns the value of a given secret.
-func GetSecret(secretName string, filter string) (string, error) {
-	out, err := get("secret", secretName, filter)
-	if err != nil {
-		return "", utils.Errorf(err, L("failed to run kubectl get secret %[1]s %[2]s"), secretName, filter)
-	}
-	decoded, err := base64.StdEncoding.DecodeString(string(out))
-	if err != nil {
-		return "", utils.Errorf(err, L("Failed to base64 decode secret %s"), secretName)
-	}
-
-	return string(decoded), nil
-}
-
-// createDockerSecret creates a secret of docker type to authenticate registries.
-func createDockerSecret(
-	namespace string,
-	name string,
-	registry string,
-	username string,
-	password string,
-	appLabel string,
-) error {
-	authString := fmt.Sprintf("%s:%s", username, password)
-	auth := base64.StdEncoding.EncodeToString([]byte(authString))
-	configjson := fmt.Sprintf(
-		`{"auths": {"%s": {"username": "%s", "password": "%s", "auth": "%s"}}}`,
-		registry, username, password, auth,
-	)
-
-	secret := core.Secret{
-		TypeMeta: meta.TypeMeta{APIVersion: "v1", Kind: "Secret"},
-		ObjectMeta: meta.ObjectMeta{
-			Namespace: namespace,
-			Name:      name,
-			Labels:    GetLabels(appLabel, ""),
-		},
-		// It seems serializing this object automatically transforms the secrets to base64.
-		Data: map[string][]byte{
-			".dockerconfigjson": []byte(configjson),
-		},
-		Type: core.SecretTypeDockerConfigJson,
-	}
-	return Apply([]runtime.Object{&secret}, fmt.Sprintf(L("failed to create the %s docker secret"), name))
-}
-
-// AddRegistry creates a secret holding the registry credentials and adds it to the helm args.
-func AddRegistry(helmArgs []string, namespace string, registry *types.Registry, appLabel string) ([]string, error) {
-	secret, err := GetRegistrySecret(namespace, registry, appLabel)
-	if secret != "" {
-		helmArgs = append(helmArgs, "--set", "registrySecret="+secret)
-	}
-	return helmArgs, err
-}
-
-var runCmdOutput = utils.RunCmdOutput
-
-// GetRegistrySecret creates a docker secret holding the registry credentials and returns the secret name.
-func GetRegistrySecret(namespace string, registry *types.Registry, appLabel string) (string, error) {
-	const secretName = "registry-credentials"
-
-	// Return the existing secret if any.
-	out, err := runCmdOutput(zerolog.DebugLevel, "kubectl", "get", "-n", namespace, "secret", secretName, "-o", "name")
-	if err == nil && strings.TrimSpace(string(out)) != "" {
-		return secretName, nil
-	}
-
-	// Create the secret if registry user and password are passed.
-	if registry.User != "" && registry.Password != "" {
-		if err := createDockerSecret(
-			namespace, secretName, registry.Host, registry.User, registry.Password, appLabel,
-		); err != nil {
-			return "", err
-		}
-		return secretName, nil
-	}
-	return "", nil
-}
-
-// GetDeploymentImagePullSecret returns the name of the image pull secret of a deployment.
-//
-// This assumes only one secret is defined on the deployment.
-func GetDeploymentImagePullSecret(namespace string, filter string) (string, error) {
-	out, err := utils.RunCmdOutput(zerolog.DebugLevel, "kubectl", "get", "deploy", "-n", namespace, filter,
-		"-o", "jsonpath={.items[*].spec.template.spec.imagePullSecrets[*].name}",
-	)
-	if err != nil {
-		return "", utils.Errorf(err, L("failed to get deployment image pull secret"))
-	}
-
-	return strings.TrimSpace(string(out)), nil
 }
