+++ uyuni-tools.changes.cbosdo.mgrpxy-nok8s
@@ -1,0 +1,1 @@
+- Remove the Kubernetes install and upgrade from mgrpxy
